# backend/openapi.yaml
openapi: 3.0.0
info:
  title: SyloPay
  description: |-
    SyloPay is a pioneering initiative that aims to establish a native and blockchain-agnostic Buy Now, Pay Later (BNPL) infrastructure on the Stellar blockchain. 
  version: 1.0.0
servers:
  - url: http://localhost:3000

paths:
  /health:
    get:
      summary: General Health Check
      description: Checks the health of the API, database, and connection to Stellar.
      tags:
        - Health
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/stellar/health:
    get:
      summary: Stellar Health Check
      description: Specifically checks the health of and connectivity to Horizon, Stellar's API server.
      tags:
        - Stellar
      responses:
        '200':
          description: Connection to Stellar successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StellarHealth'
        '500':
          description: Failed to connect to Stellar service.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/stellar/create-account:
    post:
      summary: Create an Account on Stellar
      description: Create a new account on the Stellar testnet and fund it with XLM to activate it. The secret key is only returned in the development environment.
      tags:
        - Stellar
      responses:
        '200':
          description: Account created and funded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAccountResponse'
        '500':
          description: Failed to create account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/stellar/account/{publicKey}:
    get:
      summary: Check Stellar Account.
      description: Returns the details of a specific account on the Stellar network, including its XLM balance.
      tags:
        - Stellar
      parameters:
        - name: publicKey
          in: path
          required: true
          description: The public key (address) of the Stellar account to query.
          schema:
            type: string
          example: 'GCB...56N'
      responses:
        '200':
          description: Account details found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDetailsResponse'
        '404':
          description: Account not found on the network.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/quotation:
    post:
      summary: Generate Payment Quote
      description: Generates installment options for a given amount.
      tags:
        - Payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuotationRequest'
      responses:
        '200':
          description: Quote generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationResponse'
        '400':
          description: Requisição inválida, 'amount' é obrigatório.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error when generating the quote.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/contract:
    post:
      summary: Create Payment Contract
      description: Creates a tokenized "contract" by recording transaction details in the memo field of a Stellar transaction.
      tags:
        - Payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractRequest'
      responses:
        '200':
          description: Contract successfully created and registered on the blockchain.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateContractResponse'
        '400':
          description: Invalid request, missing required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to create contract on Stellar.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message.
          example: 'Account not found'

    HealthStatus:
      type: object
      properties:
        status:
          type: string
          example: 'healthy'
        timestamp:
          type: string
          format: date-time
        services:
          type: object
          properties:
            api:
              type: string
              example: 'healthy'
            stellar:
              type: string
              example: 'healthy'
            database:
              type: string
              example: 'healthy'
        stellar:
          $ref: '#/components/schemas/StellarHealth'

    StellarHealth:
      type: object
      properties:
        connected:
          type: boolean
          example: true
        horizonUrl:
          type: string
          example: 'https://horizon-testnet.stellar.org'
        network:
          type: string
          example: 'TESTNET'
        latestLedger:
          type: number
          example: 1234567

    CreateAccountResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        account:
          type: object
          properties:
            publicKey:
              type: string
              example: 'GDA...Q5Y'
            secretKey:
              type: string
              description: "SECRET KEY - Returned only in development environment."
              example: 'SDC...34M'
        explorerUrl:
          type: string
          example: 'https://stellar.expert/explorer/testnet/account/GDA...Q5Y'
    
    AccountDetailsResponse:
      type: object
      properties:
        publicKey:
          type: string
          example: 'GDA...Q5Y'
        balance:
          type: array
          items:
            type: object
            properties:
              balance:
                type: string
                example: "9999.9999000"
              asset_type:
                type: string
                example: "native"
        sequence:
          type: string
          example: "1029384756"
        explorerUrl:
          type: string
          example: 'https://stellar.expert/explorer/testnet/account/GDA...Q5Y'

    QuotationRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: float
          description: The total value of the quote.
          example: 150.50
        installments:
          type: integer
          description: The default number of installments (optional).
          example: 3

    QuotationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        originalAmount:
          type: number
          example: 150.50
        options:
          type: array
          items:
            type: object
            properties:
              installmentsCount:
                type: integer
                example: 3
              installmentAmount:
                type: string
                example: "50.1666667"
              totalAmount:
                type: number
                example: 150.50
              description:
                type: string
                example: "3x de 50.1666667 XLM"
        currency:
          type: string
          example: "XLM"
        generatedAt:
          type: string
          format: date-time

    CreateContractRequest:
      type: object
      required:
        - merchantPublicKey
        - customerPublicKey
        - totalAmount
        - installmentsCount
      properties:
        merchantPublicKey:
          type: string
          description: Merchant account public key.
          example: 'GCM...ABC'
        customerPublicKey:
          type: string
          description: Customer account public key.
          example: 'GDX...XYZ'
        totalAmount:
          type: number
          format: float
          description: Total value of the contract.
          example: 200
        installmentsCount:
          type: integer
          description: Agreed number of installments.
          example: 4

    CreateContractResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        contract:
          type: object
          properties:
            id:
              type: string
              description: A unique ID for the contract (generated internally).
              example: 'contract_12345'
            stellarTxHash:
              type: string
              description: The transaction hash on the Stellar blockchain.
              example: 'aa16...c2e'
            explorerUrl:
              type: string
              description: Link para visualizar a transação no explorador de blocos.
              example: 'https://stellar.expert/explorer/testnet/tx/aa16...c2e'
            status:
              type: string
              example: 'created'